PDF's are no longer ignored
I want to be able to read whitepapers and books shared on this repo. Binary files are not really good to commit to revision control, but I find it handy.

This means the LaTex generated PDF file is also tracked. I am not sure if this is good or not.